YAML version of Ground and 3 Day Select rates

+++ Website/root/models.py
@@ -22,7 +22,7 @@ class Rate(models.Model):
     weight = models.IntegerField()
     service_level = models.ForeignKey(ServiceLevel)
     zone_number = models.CharField(max_length=3)
-    price = models.IntegerField()
+    price = models.FloatField()
 
     def __unicode__(self):
         return "Weight: " + str(self.weight) + ", Service level: " + self.service_level.name + ", Zone: " + str(self.zone_number)
@@ -33,6 +33,6 @@ class Shipment(models.Model):
     service_level = models.ForeignKey(ServiceLevel)
     weight = models.IntegerField()
     zone_number = models.CharField(max_length=3)
-    price = models.IntegerField()
+    price = models.FloatField()
